rmv stochastic charging
