+++ src/main/scala/edu/ie3/mobsim/MobilitySimulator.scala
@@ -17,7 +17,7 @@ import edu.ie3.mobsim.exceptions.{
 import edu.ie3.mobsim.io.geodata.PoiEnums.CategoricalLocationDictionary
 import edu.ie3.mobsim.io.geodata.{PoiUtils, PointOfInterest}
 import edu.ie3.mobsim.io.probabilities._
-import edu.ie3.mobsim.model.ChargingBehavior.chooseChargingStation
+import edu.ie3.mobsim.model.ChargingStation.chooseChargingStation
 import edu.ie3.mobsim.model.TripSimulation.simulateNextTrip
 import edu.ie3.mobsim.model.{ChargingStation, ElectricVehicle, EvType}
 import edu.ie3.mobsim.utils.{IoUtils, PathsAndSources}
@@ -395,45 +395,55 @@ final class MobilitySimulator(
       availableChargingPoints: Map[UUID, Int],
       maxDistance: ComparableQuantity[Length]
   ): Option[Movement] = {
-    val (chosenCsOpt, updatedEvOpt) = chooseChargingStation(
-      ev,
-      pricesAtChargingStation,
-      availableChargingPoints,
-      MobilitySimulator.seed,
-      maxDistance
-    )
+    val staysLongEnough = ev.parkingTimeStart
+      .until(ev.departureTime, ChronoUnit.MINUTES) >= 15
+    if (staysLongEnough) {
+      val (chosenCsOpt, updatedEvOpt) = chooseChargingStation(
+        ev,
+        pricesAtChargingStation,
+        availableChargingPoints,
+        MobilitySimulator.seed,
+        maxDistance
+      )
 
-    chosenCsOpt
-      .map { cs =>
-        val availableChargingPointsAtStation =
-          availableChargingPoints.getOrElse(cs, 0)
-        if (availableChargingPointsAtStation > 0) {
+      chosenCsOpt
+        .map { cs =>
+          val availableChargingPointsAtStation =
+            availableChargingPoints.getOrElse(cs, 0)
+          if (availableChargingPointsAtStation > 0) {
 
-          val updatedEv = updatedEvOpt
-            .getOrElse(ev)
-            .setChargingAtSimona()
-            .setChosenChargingStation(Some(cs))
+            val updatedEv = updatedEvOpt
+              .getOrElse(ev)
+              .setChargingAtSimona()
+              .setChosenChargingStation(Some(cs))
 
+            logger.debug(
+              s"${ev.getId} starts charging at $cs."
+            )
+
+            Some(Movement(cs, updatedEv))
+          } else {
+            logger.debug(
+              s"${ev.getId} could not be charged at destination ${ev.destinationPoi} " +
+                s"(${ev.getDestinationPoiType}) because all charging points " +
+                s"at $cs were taken."
+            )
+            None
+          }
+        }
+        .getOrElse {
           logger.debug(
-            s"${ev.getId} starts charging at $cs."
-          )
-
-          Some(Movement(cs, updatedEv))
-        } else {
-          logger.debug(
-            s"${ev.getId} could not be charged at destination ${ev.destinationPoi} " +
-              s"(${ev.getDestinationPoiType}) because all charging points " +
-              s"at $cs were taken."
+            s"${ev.getId} parks but does not charge."
           )
           None
         }
-      }
-      .getOrElse {
-        logger.debug(
-          s"${ev.getId} parks but does not charge."
-        )
-        None
-      }
+    } else {
+      logger.debug(
+        s"s${ev.id} parks but does not charge, since parking time is below 15 minutes."
+      )
+      None
+    }
+
   }
 
   /** Update and simulate EVs which are ending their parking at current time.
@@ -691,7 +701,7 @@ object MobilitySimulator
       config.mobsim.simulation.targetHomeChargingShare
     val start = System.currentTimeMillis()
     logger.info(
-      s"Creating $numberOfEvsInArea evs with a targeted home charging share of ${"%.2f"
+      s"Creating evs with a targeted home charging share of ${"%.2f"
           .format(targetShareOfHomeCharging * 100)} %."
     )
