+++ src/test/scala/edu/ie3/mobsim/io/geodata/PoiTestData.scala
@@ -136,7 +136,7 @@ trait PoiTestData {
     Map.empty[ChargingStation, ComparableQuantity[Length]]
   )
 
-  protected val other_shopPoi: PointOfInterest = PointOfInterest(
+  protected val otherShopPoi: PointOfInterest = PointOfInterest(
     UUID.fromString("dda899fb-e085-479d-b3a9-cba594770a2b"),
     "other_shop_251193227",
     CategoricalLocationDictionary.OTHER_SHOP,
@@ -202,7 +202,7 @@ trait PoiTestData {
     )
   )
 
-  protected val charging_hub_townPoi: PointOfInterest = PointOfInterest(
+  protected val chargingHubTownPoi: PointOfInterest = PointOfInterest(
     UUID.fromString("4df0614d-0c01-4b31-af94-804e299f2686"),
     "charging_hub_town_261344967",
     CategoricalLocationDictionary.CHARGING_HUB_TOWN,
@@ -210,7 +210,7 @@ trait PoiTestData {
     129.211731553507,
     Map(cs4 -> Quantities.getQuantity(0, METRE))
   )
-  protected val charging_hub_highwayPoi: PointOfInterest = PointOfInterest(
+  protected val chargingHubHighwayPoi: PointOfInterest = PointOfInterest(
     UUID.fromString("3ddc93c7-77fc-4187-be68-833b3db39809"),
     "charging_hub_highway_261347306",
     CategoricalLocationDictionary.CHARGING_HUB_HIGHWAY,
@@ -225,7 +225,7 @@ trait PoiTestData {
     bbpgPoi,
     culturePoi,
     medicinalPoi,
-    other_shopPoi,
+    otherShopPoi,
     religiousPoi,
     restaurantPoi,
     servicePoi,
