+++ src/test/scala/edu/ie3/mobsim/model/ChargingStationTestData.scala
@@ -6,11 +6,13 @@
 
 package edu.ie3.mobsim.model
 
+import edu.ie3.mobsim.io.geodata.PoiEnums.PoiTypeDictionary
+
 import java.time.ZonedDateTime
 import java.util.UUID
 import scala.util.Random
 
-trait ChargingBehaviorTestData extends TripSimulationTestData {
+trait ChargingStationTestData extends TripSimulationTestData {
 
   protected val evLowSoC: ElectricVehicle = ev1.copyWith(
     zero,
@@ -21,14 +23,14 @@ trait ChargingBehaviorTestData extends TripSimulationTestData {
 
   protected val evAtChargingHub: ElectricVehicle = ev2.copyWith(
     half,
-    destinationPoi = charging_hub_townPoi,
+    destinationPoi = chargingHubTownPoi,
     parkingTimeStart = ZonedDateTime.now(),
     departureTime = ZonedDateTime.now().plusHours(1)
   )
 
   protected val evNextTrip: ElectricVehicle = ev3.copyWith(
     ev3.getEStorage,
-    destinationPoi = other_shopPoi,
+    destinationPoi = otherShopPoi,
     parkingTimeStart = ZonedDateTime.now(),
     departureTime = ZonedDateTime.now().plusHours(1)
   )
