+++ src/test/scala/edu/ie3/mobsim/utils/IoUtilsTestData.scala
@@ -8,14 +8,14 @@ package edu.ie3.mobsim.utils
 
 import edu.ie3.mobsim.io.geodata.PoiEnums.CategoricalLocationDictionary
 import edu.ie3.mobsim.io.geodata.PointOfInterest
-import edu.ie3.mobsim.model.{ChargingBehaviorTestData, ElectricVehicle}
+import edu.ie3.mobsim.model.{ChargingStationTestData, ElectricVehicle}
 
 import java.io.File
 import java.nio.file.Files
 import java.time.ZonedDateTime
 import java.util.UUID
 
-trait IoUtilsTestData extends ChargingBehaviorTestData {
+trait IoUtilsTestData extends ChargingStationTestData {
   protected val outputFileDir: File =
     Files.createTempDirectory("io-utils-test").toFile
 
@@ -38,7 +38,7 @@ trait IoUtilsTestData extends ChargingBehaviorTestData {
 
   protected val poiMap
       : Map[CategoricalLocationDictionary.Value, Set[PointOfInterest]] = {
-    val poiSet: Set[PointOfInterest] = Seq(charging_hub_townPoi).toSet
+    val poiSet: Set[PointOfInterest] = Seq(chargingHubTownPoi).toSet
 
     poiSet.groupBy(_.categoricalLocation).map { case (catLoc, poi) =>
       catLoc -> Set.from(poi)
